+++ nexus/extensions/audio/dsp_video_encoder/include/priv/nexus_audio_dsp_video_encoder.h
@@ -1,5 +1,5 @@
 /******************************************************************************
- * Copyright (C) 2017 Broadcom.  The term "Broadcom" refers to Broadcom Limited and/or its subsidiaries.
+ * Copyright (C) 2017-2018 Broadcom.  The term "Broadcom" refers to Broadcom Inc. and/or its subsidiaries.
  *
  * This program is the proprietary software of Broadcom and/or its licensors,
  * and may only be used, duplicated, modified or distributed pursuant to the terms and
@@ -105,7 +105,7 @@ typedef struct NEXUS_DspVideoEncoderDescriptor
 
    /* Buffer Parameters */
    unsigned offset; /* offset of frame data from frame buffer base address (in bytes) */
-   size_t length;   /* 0 if fragment is empty, e.g. for EOS entry (in bytes) */
+   unsigned length; /* 0 if fragment is empty, e.g. for EOS entry (in bytes) */
    unsigned decoderIndex;
    unsigned unused;
    uint32_t videoFlags;
